# apps/rights-web/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rights Institute API",
    "version": "1.0.0",
    "description": "API for Rights Institute - Authentication and User Management",
    "contact": {
      "name": "Rights Institute",
      "url": "https://rights.institute"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://rights.institute",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8787",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/auth/get-session": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current session",
        "description": "Retrieve the current user's session information",
        "operationId": "getSession",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Session information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "example": "https://example.com/avatar.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-01T00:00:00.000Z"
          }
        }
      },
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "session_123e4567"
          },
          "userId": {
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-01-08T00:00:00.000Z"
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "session": {
            "$ref": "#/components/schemas/Session"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Authentication failed"
          },
          "message": {
            "type": "string",
            "example": "Invalid email or password"
          }
        }
      }
    }
  }
}